+++ packages/bridge/Editor/Bridge/MainThreadDispatcher.cs
@@ -1,6 +1,9 @@
 using System;
 using System.Collections.Concurrent;
+using System.Threading;
+using System.Threading.Tasks;
 using UnityEditor;
+using UnityEngine;
 
 namespace UnityAgentBridge
 {
@@ -33,6 +36,28 @@ namespace UnityAgentBridge
         public static void Enqueue(Action action)
         {
             _queue.Enqueue(action);
+        }
+
+        public static Task<T> EnqueueAsync<T>(Func<T> action, int timeoutMs)
+        {
+            var tcs = new TaskCompletionSource<T>(TaskCreationOptions.RunContinuationsAsynchronously);
+
+            _queue.Enqueue(() =>
+            {
+                try
+                {
+                    tcs.TrySetResult(action());
+                }
+                catch (Exception e)
+                {
+                    tcs.TrySetException(e);
+                }
+            });
+
+            var timer = new Timer(_ => tcs.TrySetException(new TimeoutException()), null, timeoutMs, Timeout.Infinite);
+            tcs.Task.ContinueWith(_ => timer.Dispose());
+
+            return tcs.Task;
         }
     }
 }